refactor(caps): split monolithic caps module into focused submodules
