+++ src/openhuman/tinyflows/caps/mod.rs
@@ -61,7 +61,24 @@ pub(crate) use crate::openhuman::composio::catalog::{
 #[cfg(test)]
 pub(crate) use crate::openhuman::composio::catalog::{seed_live_catalog_cache, seed_probe_cache};
 
+mod tier;
+mod llm;
+mod mocks;
+mod http;
+mod code;
+mod state;
+mod resolver;
 pub(crate) mod tools;
+
+// Glob re-exports keep every existing path resolving after the split:
+// `caps::X` from `flows/`, `super::super::X` from `caps::tools::*`.
+pub(crate) use tier::*;
+pub(crate) use llm::*;
+pub(crate) use mocks::*;
+pub(crate) use http::*;
+pub(crate) use code::*;
+pub(crate) use state::*;
+pub(crate) use resolver::*;
 
 pub(crate) use tools::NATIVE_TOOL_PREFIX;
 
@@ -107,190 +124,6 @@ fn model_response_to_completion_value(
     })
 }
 
-/// Hard autonomy-tier gate for an *acting* flow node (Phase 2).
-///
-/// A flow run scopes a `TrustedAutomation { Workflow }` origin, but the acting
-/// power of a run is still bounded by the user's `[autonomy]` tier — the same
-/// [`SecurityPolicy`] the agent tool-loop honors (`SecurityPolicy::from_config`
-/// off the `[autonomy]` block). Before an `http_request` (Network-class) or
-/// `code` (Write-class) node dispatches, we consult
-/// [`SecurityPolicy::gate_decision`] for that node's [`CommandClass`] and refuse
-/// outright when the tier `Block`s it — mirroring how `curl`/`shell` acting
-/// tools gate (`policy.gate_decision(CommandClass::Network)`), so a read-only
-/// run can never reach the network or run arbitrary code.
-///
-/// `Allow`/`Prompt` return `Ok(decision)`: this function only enforces the
-/// non-negotiable `Block` floor itself. The caller uses the returned
-/// [`GateDecision`] to drive [`gate_call_for_tier`] immediately after, which is
-/// what actually performs the `Prompt` round-trip (see that function's doc for
-/// why this is not automatic — a saved workflow's own `require_approval` flag
-/// would otherwise silently override the tier's `Prompt` decision). The error
-/// is prefixed with [`POLICY_BLOCKED_MARKER`] so the harness's repeated-failure
-/// middleware recognizes it as a permanent, don't-retry refusal.
-fn enforce_node_tier_gate(
-    security: &SecurityPolicy,
-    class: CommandClass,
-    node: &str,
-) -> Result<GateDecision> {
-    let decision = security.gate_decision(class);
-    tracing::debug!(
-        target: "flows",
-        node,
-        ?class,
-        ?decision,
-        tier = ?security.autonomy,
-        "[flows] node tier gate: evaluating autonomy-tier decision"
-    );
-    if decision == GateDecision::Block {
-        tracing::warn!(
-            target: "flows",
-            node,
-            ?class,
-            tier = ?security.autonomy,
-            "[flows] node tier gate: BLOCKED by autonomy tier — refusing before dispatch"
-        );
-        return Err(EngineError::Capability(format!(
-            "{POLICY_BLOCKED_MARKER} flows {node} node is not permitted under the current \
-             autonomy tier ({:?}): {class:?}-class actions are blocked. Raise the [autonomy] \
-             tier to run this node.",
-            security.autonomy
-        )));
-    }
-    Ok(decision)
-}
-
-/// Dispatches to the process-global [`ApprovalGate`](crate::openhuman::approval::ApprovalGate),
-/// escalating a `Prompt`-tier decision into a forced human-in-the-loop round
-/// trip regardless of the running flow's own `require_approval` toggle.
-///
-/// **Why this is needed (Codex P1 finding):** `ApprovalGate::intercept_audited`
-/// branches on the scoped [`AgentTurnOrigin`](crate::openhuman::agent::turn_origin::AgentTurnOrigin) —
-/// for a `TrustedAutomation { source: Workflow { require_approval: false }, .. }`
-/// origin (the default for every saved flow unless the author opts in) it
-/// returns `Allow` unconditionally, the same pre-declared-trust-root shortcut a
-/// user-authorized cron job gets. That shortcut is correct when the node's
-/// autonomy-tier decision was itself `Allow`, but it silently defeats a
-/// Supervised-tier `Prompt` decision: without this escalation, a Supervised
-/// user's `http_request`/`code` node would run unattended purely because the
-/// flow's `require_approval` defaults to `false` — the tier's "ask me" was
-/// never actually enforced.
-///
-/// When `tier_decision` is [`GateDecision::Prompt`] and the current origin is a
-/// `Workflow { require_approval: false }` trust root, this scopes a *for this
-/// call only* `Workflow { require_approval: true }` origin around
-/// `intercept_audited`, forcing the real parking/HITL flow. `GateDecision::Allow`
-/// (and any other origin shape) passes through unchanged — existing behavior.
-async fn gate_call_for_tier(
-    tier_decision: GateDecision,
-    tool_name: &str,
-    action_summary: &str,
-    args_redacted: Value,
-) -> (crate::openhuman::approval::GateOutcome, Option<String>) {
-    use crate::openhuman::agent::turn_origin;
-
-    let Some(gate) = crate::openhuman::approval::ApprovalGate::try_global() else {
-        return (crate::openhuman::approval::GateOutcome::Allow, None);
-    };
-
-    match escalated_origin_for_prompt(tier_decision, turn_origin::current()) {
-        Some(escalated) => {
-            tracing::debug!(
-                target: "flows",
-                tool_name,
-                "[flows] node tier gate: tier decision is Prompt — escalating this dispatch to a \
-                 forced approval round-trip regardless of the flow's require_approval toggle"
-            );
-            turn_origin::with_origin(
-                escalated,
-                gate.intercept_audited(tool_name, action_summary, args_redacted),
-            )
-            .await
-        }
-        None => {
-            gate.intercept_audited(tool_name, action_summary, args_redacted)
-                .await
-        }
-    }
-}
-
-/// Pure decision core of [`gate_call_for_tier`]: when `tier_decision` is
-/// [`GateDecision::Prompt`] and `origin` is a `Workflow { require_approval:
-/// false }` trust root, returns a clone of that origin with `require_approval`
-/// flipped to `true` (the forced escalation). Otherwise returns `None` — the
-/// caller then dispatches through the unmodified origin, matching prior
-/// behavior. Split out as a free function over plain values (no gate, no
-/// task-local read) so the escalation policy is unit-testable without a live
-/// `ApprovalGate`.
-fn escalated_origin_for_prompt(
-    tier_decision: GateDecision,
-    origin: Option<crate::openhuman::agent::turn_origin::AgentTurnOrigin>,
-) -> Option<crate::openhuman::agent::turn_origin::AgentTurnOrigin> {
-    use crate::openhuman::agent::turn_origin::{AgentTurnOrigin, TrustedAutomationSource};
-
-    if tier_decision != GateDecision::Prompt {
-        return None;
-    }
-    match origin {
-        Some(AgentTurnOrigin::TrustedAutomation {
-            job_id,
-            source:
-                TrustedAutomationSource::Workflow {
-                    require_approval: false,
-                },
-        }) => Some(AgentTurnOrigin::TrustedAutomation {
-            job_id,
-            source: TrustedAutomationSource::Workflow {
-                require_approval: true,
-            },
-        }),
-        _ => None,
-    }
-}
-
-/// Pure decision core of the nested agent-node harness escalation (issue
-/// #4595): when the flow run's origin is a `Workflow { require_approval: false }`
-/// trust root, returns a clone with `require_approval` flipped to `true` so the
-/// [`ApprovalGate`](crate::openhuman::approval::ApprovalGate)'s pre-declared-
-/// action shortcut (`gate.rs::intercept_audited`, `Workflow { require_approval:
-/// false }` → `Allow` without prompt) does NOT apply to tool calls the nested
-/// harness picks at runtime.
-///
-/// **Why this is different from [`escalated_origin_for_prompt`].** That helper
-/// escalates a *single* flow-node acting tool dispatch when the tier decision
-/// is `Prompt`. This helper escalates the *entire nested harness turn*
-/// unconditionally, because the flow author never pre-declared which tools the
-/// referenced agent's LLM will pick — the graph only names the `agent_ref`, and
-/// the definition's `ToolScope` is the runtime pool. So the "trust root =
-/// static action" invariant that justifies the `intercept_audited` shortcut
-/// simply doesn't hold across the `Agent::run_single` boundary.
-///
-/// `Workflow { require_approval: true }` passes through unchanged (already
-/// user-forced HITL); other origins pass through unchanged (Cron / Web chat
-/// / etc. don't route through this call site today, but if they ever do the
-/// shortcut is safe or already covered by that origin's own gate branch).
-/// Split out as a free function over plain values so the escalation policy is
-/// unit-testable without a live `ApprovalGate`.
-fn escalated_origin_for_nested_harness(
-    origin: Option<crate::openhuman::agent::turn_origin::AgentTurnOrigin>,
-) -> Option<crate::openhuman::agent::turn_origin::AgentTurnOrigin> {
-    use crate::openhuman::agent::turn_origin::{AgentTurnOrigin, TrustedAutomationSource};
-
-    match origin {
-        Some(AgentTurnOrigin::TrustedAutomation {
-            job_id,
-            source:
-                TrustedAutomationSource::Workflow {
-                    require_approval: false,
-                },
-        }) => Some(AgentTurnOrigin::TrustedAutomation {
-            job_id,
-            source: TrustedAutomationSource::Workflow {
-                require_approval: true,
-            },
-        }),
-        _ => None,
-    }
-}
 
 /// Cap on the serialized `input_context` block size (bytes of the pretty-
 /// printed JSON) before truncation. Keeps a huge upstream payload (e.g. a
@@ -544,133 +377,6 @@ fn resolve_completion_model(node_model: Option<&str>, resolved_model: String) ->
     }
 }
 
-/// [`LlmProvider`] adapter over OpenHuman's inference stack
-/// (`src/openhuman/inference/provider/`).
-///
-/// The `agent` node is single-completion in tinyflows 0.2 (no tool-calling
-/// loop, no sub-ports), so `complete` performs exactly one `provider.chat`
-/// call and returns its result — no agent loop is driven here.
-///
-/// **Structured output**: when the node requested it (an
-/// `output_parser.schema` or `response_format: "json"` in the config), the
-/// completion text is parsed as JSON and the **parsed object** is returned as
-/// the response value; otherwise the `{text: "..."}` shape is returned. Either
-/// way the tinyflows `agent` node wraps this in its stable output **envelope**
-/// `{ json, text, raw }`, so a downstream node binds `=item.json.<field>` for
-/// structured output or `=item.text` for prose (or
-/// `=nodes.<agent_id>.item.json.<field>` across nodes) — the parsed-vs-`{text}`
-/// shape is no longer visible to consumers. A completion that doesn't parse
-/// still lets the agent node's `output_parser` sub-port coerce it via the
-/// schema auto-fix path before enveloping.
-pub struct OpenHumanLlm {
-    pub config: Arc<Config>,
-}
-
-#[async_trait]
-impl LlmProvider for OpenHumanLlm {
-    async fn complete(&self, request: Value, conn: Option<&str>) -> Result<Value> {
-        if let Some(c) = conn {
-            // B1 does not resolve `connection_ref` to a specific BYOK account —
-            // `create_chat_provider` picks the configured provider for `role`.
-            tracing::debug!(target: "flows", conn = %c, "[flows] llm conn (not resolved in B1)");
-        }
-
-        let role = request
-            .get("role")
-            .and_then(Value::as_str)
-            .unwrap_or("summarization");
-
-        // Per-node model selection: an `agent` node may pin a **managed tier**
-        // (`config.model = "reasoning-v1"` / `"chat-v1"`, or a `hint:*` alias).
-        // Map that tier back to the workload role whose provider serves it so
-        // the completion routes to that tier on the managed backend (or the
-        // role's BYOK model) instead of the node's default `role`. Unknown /
-        // absent model strings leave the role untouched. `config.model` is
-        // trusted node config, never model output.
-        let node_model = request
-            .get("model")
-            .and_then(Value::as_str)
-            .map(str::trim)
-            .filter(|s| !s.is_empty());
-        let role = match node_model {
-            Some(model) => {
-                let mapped = role_for_model_tier(model);
-                tracing::debug!(
-                    target: "flows",
-                    node_model = model,
-                    mapped_role = mapped,
-                    "[flows] llm.complete: node pinned a model tier — routing by mapped role"
-                );
-                mapped
-            }
-            None => role,
-        };
-        let temperature = request
-            .get("temperature")
-            .and_then(Value::as_f64)
-            .unwrap_or(0.7);
-        let max_tokens = request
-            .get("max_tokens")
-            .and_then(Value::as_u64)
-            .and_then(|n| u32::try_from(n).ok());
-
-        let structured = structured_output_requested(&request);
-        let messages = build_completion_messages(&request);
-
-        tracing::debug!(
-            target: "flows",
-            role,
-            message_count = messages.len(),
-            structured,
-            "[flows] llm.complete: dispatching agent-node completion"
-        );
-
-        let (chat_model, model) = create_chat_model_with_model_id(role, &self.config, temperature)
-            .map_err(|e| EngineError::Capability(e.to_string()))?;
-        // `create_chat_provider` handed back the role's default model. If the node
-        // pinned a raw/BYOK id, forward it verbatim instead (issue #4598).
-        let model = resolve_completion_model(node_model, model);
-
-        let mut model_request = ModelRequest::new(
-            messages
-                .iter()
-                .map(crate::openhuman::tinyagents::chat_message_to_message)
-                .collect(),
-        )
-        .with_model(model.clone())
-        .with_temperature(temperature);
-        if let Some(max_tokens) = max_tokens {
-            model_request = model_request.with_max_tokens(max_tokens);
-        }
-        let response = chat_model
-            .invoke(&(), model_request)
-            .await
-            .map_err(|e| EngineError::Capability(e.to_string()))?;
-
-        // Structured mode: surface the parsed object itself so downstream
-        // `=item.<field>` / `=nodes.<id>.item.<field>` bindings work. The
-        // agent node's output_parser sub-port then validates it against the
-        // configured schema (and auto-fixes when it doesn't parse here).
-        if structured {
-            let text = response.text();
-            if let Some(parsed) = parse_llm_json(&text) {
-                tracing::debug!(
-                    target: "flows",
-                    "[flows] llm.complete: structured output parsed from completion text"
-                );
-                return Ok(parsed);
-            }
-            tracing::warn!(
-                target: "flows",
-                "[flows] llm.complete: structured output requested but the completion did not \
-                 parse as JSON — falling back to the {{text}} shape (the output_parser sub-port \
-                 may still coerce it)"
-            );
-        }
-
-        Ok(model_response_to_completion_value(&response))
-    }
-}
 
 /// [`AgentRunner`] backing an `agent` node's `agent_ref`. It runs the selected
 /// agent kind by one of two paths, chosen by [`route_for_agent_ref`]:
@@ -1334,186 +1040,6 @@ fn prepend_system_message(request: &mut Value, system_prompt: &str) {
     }
 }
 
-/// A **dry-run-only** [`AgentRunner`] mock that, unlike the vendored crate's
-/// `tinyflows::caps::mock::MockAgentRunner`, respects an `agent` node's
-/// `config.output_parser.schema` when synthesizing its echo response.
-///
-/// `DryRunWorkflowTool` (`flows::builder_tools`) wires this in place of the
-/// vendored `MockAgentRunner` so its null-resolution check (every `tool_call`
-/// arg that resolves to `null`) doesn't **false-positive** on a CORRECTLY-built
-/// agent node. Without it: the vendored `MockAgentRunner` always echoes
-/// `{ agent, request, connection }` regardless of schema, and the vendored
-/// `agent` node's output-parser sub-port (`tinyflows::nodes::integration::schema`)
-/// then fails that shape against ANY declared schema (no field matches) and
-/// falls to a one-shot LLM auto-fix that the sandbox's plain `MockLlm` also
-/// can't satisfy — so the whole dry run would error out even for a workflow a
-/// real run (via [`OpenHumanAgentRunner`], whose completion the same sub-port
-/// validates/repairs against the schema) would execute cleanly.
-///
-/// When `request` (the resolved node config `run_agent` receives — see
-/// [`AgentRunner::run_agent`]) carries a non-null `output_parser.schema`
-/// describing an object with `properties`, returns an object with every
-/// declared property present, populated with a type-appropriate placeholder
-/// (`string` → `""`, `number`/`integer` → `0`, `boolean` → `false`, `object` →
-/// `{}`, `array` → `[]`, anything else → `null`; a property with a non-empty
-/// `enum` gets its FIRST allowed value instead — see [`placeholder_for_type`])
-/// — enough to satisfy the vendored validator's `type`/`required`/`enum`
-/// checks (see `tinyflows::nodes::integration::schema::validate`) without a
-/// real model call. With no schema, mirrors the vendored `MockAgentRunner`'s
-/// default echo shape so dry-run behavior for schema-less agents is unchanged.
-#[derive(Debug, Default, Clone)]
-pub struct SchemaAwareMockAgentRunner;
-
-#[async_trait]
-impl AgentRunner for SchemaAwareMockAgentRunner {
-    async fn run_agent(
-        &self,
-        agent_ref: &str,
-        request: Value,
-        conn: Option<&str>,
-    ) -> Result<Value> {
-        let schema = request
-            .get("output_parser")
-            .and_then(|parser| parser.get("schema"))
-            .filter(|schema| !schema.is_null());
-        match schema {
-            Some(schema) => {
-                let placeholder = placeholder_for_schema(schema);
-                tracing::debug!(
-                    target: "flows",
-                    agent_ref,
-                    "[flows] dry_run: schema-aware mock agent synthesized a placeholder \
-                     matching output_parser.schema"
-                );
-                Ok(placeholder)
-            }
-            None => {
-                tracing::debug!(
-                    target: "flows",
-                    agent_ref,
-                    "[flows] dry_run: schema-aware mock agent has no output_parser.schema — \
-                     mirroring the vendored MockAgentRunner echo shape"
-                );
-                Ok(json!({ "agent": agent_ref, "request": request, "connection": conn }))
-            }
-        }
-    }
-}
-
-/// A **dry-run-only** [`LlmProvider`] mock that, unlike the vendored crate's
-/// `tinyflows::caps::mock::MockLlm`, respects an `agent` node's
-/// `config.output_parser.schema` when synthesizing its completion.
-///
-/// This closes the OTHER half of the same gap [`SchemaAwareMockAgentRunner`]
-/// closes. The vendored `agent` node only routes to an [`AgentRunner`] when the
-/// node carries a **non-empty `agent_ref`** AND the host wired an agent registry
-/// (`vendor/tinyflows/src/nodes/integration/agent.rs`, `run_turn`:
-/// `(Some(agent_ref), Some(runner)) => runner.run_agent(...)`); **every other
-/// case** — and builder-generated agent nodes carry NO `agent_ref` — falls back
-/// to `ctx.caps.llm.complete(cfg.clone(), conn)`. So in the sandbox those plain
-/// agent nodes never reach `SchemaAwareMockAgentRunner` at all: they hit the
-/// `llm` slot, which (with the vendored `MockLlm`) echoes
-/// `{ "completion": <config>, "connection": <conn> }`. The agent node's
-/// output-parser sub-port then validates that echo against the declared schema
-/// (`schema::parse_and_validate` — it validates the WHOLE completion value, not
-/// a `.text` field), no field matches, and it falls to a one-shot LLM auto-fix
-/// that the same `MockLlm` also can't satisfy — so the dry run errors with
-/// `output_parser: value failed schema validation after auto-fix: missing
-/// required property ...` even for a workflow a real run would execute cleanly.
-/// This false-failure burned many dry-run cycles for correctly-built graphs.
-///
-/// When `request` (the node config the node hands to `complete` — see the
-/// `_ => ctx.caps.llm.complete(cfg.clone(), conn)` arm above) carries a non-null
-/// `output_parser.schema`, this returns [`placeholder_for_schema`] DIRECTLY.
-/// The sub-port receives that already-schema-valid object as its `value`
-/// (`validate` returns no errors), so it returns `Ok` WITHOUT ever invoking the
-/// auto-fix LLM path — exactly the shape the vendored validator's
-/// `type`/`required`/`enum` checks accept, with no real model call. With no
-/// schema, it mirrors the vendored `MockLlm` echo shape byte-for-byte
-/// (`{ "completion": request, "connection": conn }`) so schema-less agent
-/// dry-run behavior — and downstream `=nodes.<agent>.item.json.completion...`
-/// bindings — stay identical to today.
-#[derive(Debug, Default, Clone)]
-pub struct SchemaAwareMockLlm;
-
-#[async_trait]
-impl LlmProvider for SchemaAwareMockLlm {
-    async fn complete(&self, request: Value, conn: Option<&str>) -> Result<Value> {
-        let schema = request
-            .get("output_parser")
-            .and_then(|parser| parser.get("schema"))
-            .filter(|schema| !schema.is_null());
-        match schema {
-            Some(schema) => {
-                let placeholder = placeholder_for_schema(schema);
-                tracing::debug!(
-                    target: "flows",
-                    "[flows] dry_run: schema-aware mock LLM synthesized a placeholder \
-                     matching output_parser.schema (plain agent node, no agent_ref)"
-                );
-                Ok(placeholder)
-            }
-            None => {
-                tracing::debug!(
-                    target: "flows",
-                    "[flows] dry_run: schema-aware mock LLM has no output_parser.schema — \
-                     mirroring the vendored MockLlm echo shape"
-                );
-                Ok(json!({ "completion": request, "connection": conn }))
-            }
-        }
-    }
-}
-
-/// Builds a placeholder JSON value satisfying `schema`'s `properties`/`type`
-/// constraints, for [`SchemaAwareMockAgentRunner`]. Only the shallow, top-level
-/// `properties` map is populated — enough for the minimal validator in
-/// `tinyflows::nodes::integration::schema` (`type`, `required`, `properties`);
-/// deeply-nested `required` constraints on a nested `object`/`array` property
-/// are a documented limitation (the placeholder for those is an empty `{}`/`[]`).
-fn placeholder_for_schema(schema: &Value) -> Value {
-    match schema.get("properties").and_then(Value::as_object) {
-        Some(props) => {
-            let placeholders = props
-                .iter()
-                .map(|(key, subschema)| (key.clone(), placeholder_for_type(subschema)));
-            Value::Object(placeholders.collect())
-        }
-        // No `properties` to enumerate (e.g. a bare `{"type": "array"}`
-        // schema) — fall back to a type-only placeholder for the schema itself.
-        None => placeholder_for_type(schema),
-    }
-}
-
-/// The placeholder value for one property's subschema, keyed by its
-/// declared JSON-Schema `type` (see [`placeholder_for_schema`]).
-///
-/// An `enum` constraint is honored FIRST, before falling back to the
-/// type-only placeholder: the vendored validator
-/// (`tinyflows::nodes::integration::schema::validate`) rejects any value not
-/// listed in a schema's `enum`, and a generic type placeholder (e.g. `""` for
-/// `{"type": "string", "enum": ["urgent", "normal"]}`) is essentially never
-/// one of the allowed values — that would fail the dry run even though a real
-/// agent, prompted with the schema, could easily satisfy it. The schema
-/// author's own first listed value is always allowed by construction, so it's
-/// returned as-is (whatever its JSON type).
-fn placeholder_for_type(subschema: &Value) -> Value {
-    if let Some(first_allowed) = subschema
-        .get("enum")
-        .and_then(Value::as_array)
-        .and_then(|values| values.first())
-    {
-        return first_allowed.clone();
-    }
-    match subschema.get("type").and_then(Value::as_str) {
-        Some("string") => json!(""),
-        Some("number" | "integer") => json!(0),
-        Some("boolean") => json!(false),
-        Some("object") => json!({}),
-        Some("array") => json!([]),
-        _ => Value::Null,
-    }
-}
 
 /// Parses a `"composio:<toolkit>:<connection_id>"` `connection_ref` (see the
 /// node catalog, `my_docs/ohxtf/commons/12-node-catalog-0.2.md`) and returns
@@ -2027,506 +1553,9 @@ impl ToolInvoker for OpenHumanTools {
     }
 }
 
-/// [`HttpClient`] adapter over `HttpRequestTool`
-/// (`src/openhuman/tools/impl/network/http_request.rs`). Allowlist + DNS-rebind
-/// guard live inside `execute`, so this adapter gets them for free.
-///
-/// **B2:** also routes through the OpenHuman `ApprovalGate` before dispatch
-/// (same rationale/shape as [`OpenHumanTools::invoke`] — closes the Codex P1
-/// finding that flow HTTP nodes bypassed the Network approval gate).
-///
-/// **Phase 2 — `http_cred:<name>` resolution:** a `"http_cred:<name>"`
-/// `connection_ref` is now resolved against the credentials domain's
-/// [`HttpCredentialsStore`] (encrypted-at-rest bearer/basic/header templates).
-/// The resolved auth header is injected **server-side** into the outbound
-/// request — after the approval gate has already computed its redacted audit
-/// summary — so the secret is never surfaced to the approval UI, the flow
-/// engine/graph, the node's output, or the logs (only the header *name* and
-/// scheme are logged; the value is redacted). A `connection_ref` that names an
-/// **unknown** credential fails the request closed (`EngineError::Capability`)
-/// rather than silently sending it unauthenticated.
-pub struct OpenHumanHttp {
-    pub security: Arc<SecurityPolicy>,
-    pub http_config: HttpRequestConfig,
-    pub http_creds: Arc<HttpCredentialsStore>,
-}
 
-/// Resolves an optional HTTP `connection_ref` to the stored credential to
-/// inject. Split out as a free function (over the store, not `&self`) so the
-/// resolve/fail-closed policy is unit-testable without constructing a full
-/// [`OpenHumanHttp`] adapter.
-///
-/// - `None` conn, or a `connection_ref` whose prefix isn't `http_cred:` →
-///   `Ok(None)` (no credential to inject; a non-`http_cred:` prefix is logged
-///   and ignored, matching the pre-Phase-2 behavior).
-/// - a `http_cred:<name>` naming a **known** credential → `Ok(Some(cred))`
-///   (secret-bearing — the caller injects it server-side, never logs it).
-/// - a `http_cred:<name>` naming an **unknown** credential, a malformed
-///   (empty/whitespace-only) name, or a store error → `Err` — the request
-///   must fail closed, never proceed unauthenticated. Distinguishing "no
-///   `http_cred:` prefix at all" from "`http_cred:` prefix with a malformed
-///   name" matters: [`http_cred_name`] collapses both to `None`, which would
-///   otherwise let a typo'd or data-derived empty ref (e.g. `"http_cred:"`)
-///   silently fall through to an unauthenticated request (Codex P2 finding).
-fn resolve_http_credential(
-    store: &HttpCredentialsStore,
-    conn: Option<&str>,
-) -> Result<Option<HttpCredential>> {
-    let Some(conn) = conn else {
-        return Ok(None);
-    };
-    if conn.strip_prefix("http_cred:").is_none() {
-        tracing::debug!(target: "flows", %conn, "[flows] http conn: unrecognized connection_ref prefix (expected `http_cred:<name>`) — ignoring");
-        return Ok(None);
-    }
-    let Some(name) = http_cred_name(conn) else {
-        tracing::warn!(
-            target: "flows",
-            %conn,
-            "[flows] http_request: connection_ref has the `http_cred:` prefix but no credential \
-             name — failing the request closed rather than sending it unauthenticated"
-        );
-        return Err(EngineError::Capability(format!(
-            "http_request connection_ref has a malformed http_cred name: {conn:?}"
-        )));
-    };
 
-    match store.get(name) {
-        Ok(Some(cred)) => {
-            tracing::debug!(
-                target: "flows",
-                cred = %name,
-                scheme = cred.scheme.as_str(),
-                "[flows] http_request: resolved http_cred (secret redacted)"
-            );
-            Ok(Some(cred))
-        }
-        Ok(None) => {
-            tracing::warn!(
-                target: "flows",
-                cred = %name,
-                "[flows] http_request: connection_ref names an unknown http_cred — failing the \
-                 request closed rather than sending it unauthenticated"
-            );
-            Err(EngineError::Capability(format!(
-                "http_request connection_ref names an unknown http_cred: {name}"
-            )))
-        }
-        Err(e) => {
-            tracing::error!(
-                target: "flows",
-                cred = %name,
-                error = %e,
-                "[flows] http_request: failed to resolve http_cred from the store"
-            );
-            Err(EngineError::Capability(format!(
-                "failed to resolve http_cred '{name}': {e}"
-            )))
-        }
-    }
-}
 
-/// Merges a resolved credential's auth header into the outbound `request`'s
-/// `headers` object (creating it when absent), returning the header **name**
-/// that was injected for redacted logging. The header value carries the secret
-/// and is placed only into the request handed to `HttpRequestTool` — it is
-/// never logged or returned. An explicit stored credential wins over any inline
-/// same-named header the flow author set.
-fn inject_http_credential(request: &mut Value, cred: &HttpCredential) -> Result<String> {
-    let (header_name, header_value) = cred
-        .to_header()
-        .map_err(|e| EngineError::Capability(e.to_string()))?;
-
-    let obj = request.as_object_mut().ok_or_else(|| {
-        EngineError::Capability("http_request config must be a JSON object".to_string())
-    })?;
-    let headers_entry = obj
-        .entry("headers")
-        .or_insert_with(|| Value::Object(serde_json::Map::new()));
-    // A flow author may leave `headers` unset (null) — coerce to an object so
-    // the credential still injects. A non-object, non-null `headers` is a
-    // malformed config we refuse rather than silently drop the credential.
-    if headers_entry.is_null() {
-        *headers_entry = Value::Object(serde_json::Map::new());
-    }
-    let headers_obj = headers_entry.as_object_mut().ok_or_else(|| {
-        EngineError::Capability("http_request `headers` must be a JSON object".to_string())
-    })?;
-    headers_obj.insert(header_name.clone(), Value::String(header_value));
-
-    tracing::info!(
-        target: "flows",
-        cred = %cred.name,
-        scheme = cred.scheme.as_str(),
-        header = %header_name,
-        "[flows] http_request: injected stored credential header (value redacted)"
-    );
-    Ok(header_name)
-}
-
-#[async_trait]
-impl HttpClient for OpenHumanHttp {
-    async fn request(&self, mut request: Value, conn: Option<&str>) -> Result<Value> {
-        const TOOL_NAME: &str = "flows_http_request";
-
-        // Autonomy-tier gate (Phase 2): an http_request node reaches the network,
-        // so it is Network-class. A read-only run `Block`s here and never
-        // dispatches; Supervised/Full fall through to the ApprovalGate below.
-        // `gate_call_for_tier` is what actually performs the `Prompt` round-trip
-        // — it escalates a Supervised `Prompt` decision into a forced approval
-        // regardless of the flow's own `require_approval` toggle (Codex P1).
-        let tier_decision =
-            enforce_node_tier_gate(&self.security, CommandClass::Network, "http_request")?;
-
-        // The approval gate summarizes/redacts the request BEFORE any credential
-        // is injected, so a stored secret never lands in the approval UI or
-        // audit trail. Injection happens strictly after this point.
-        let summary = crate::openhuman::approval::summarize_action(TOOL_NAME, &request);
-        let redacted = crate::openhuman::approval::redact_args(&request);
-        let (outcome, audit_id) =
-            gate_call_for_tier(tier_decision, TOOL_NAME, &summary, redacted).await;
-        if let crate::openhuman::approval::GateOutcome::Deny { reason } = outcome {
-            return Err(EngineError::Capability(reason));
-        }
-
-        // Resolve `http_cred:<name>` to a stored credential and inject its auth
-        // header server-side. An unknown name fails the request closed (see
-        // `resolve_http_credential`) — we never send it unauthenticated.
-        if let Some(cred) = resolve_http_credential(&self.http_creds, conn)? {
-            inject_http_credential(&mut request, &cred)?;
-        }
-
-        let tool = HttpRequestTool::new(
-            self.security.clone(),
-            self.http_config.allowed_domains.clone(),
-            self.http_config.max_response_size,
-            self.http_config.timeout_secs,
-        );
-
-        tracing::debug!(
-            target: "flows",
-            method = ?request.get("method"),
-            url = ?request.get("url"),
-            "[flows] http_request: dispatching outbound request"
-        );
-
-        // `request` is already `{ method, url, headers?, body? }` — the node's
-        // config is the request descriptor; `HttpRequestTool::execute` reads
-        // only those keys and ignores the rest (e.g. `connection_ref`,
-        // `on_error`), so passing the whole config through is safe.
-        let result = tool.execute(request).await;
-
-        let outcome: Result<Value> = match result {
-            Ok(result) if result.is_error => {
-                // `HttpRequestTool::execute` always returns `Ok`, using
-                // `is_error` to signal a failed request (non-2xx, DNS/allowlist
-                // rejection, timeout, …) — surface that as a capability error
-                // so the engine's `on_error`/`retry` policy can act on it.
-                Err(EngineError::Capability(result.text()))
-            }
-            Ok(result) => Ok(json!({ "text": result.text() })),
-            Err(e) => Err(EngineError::Capability(e.to_string())),
-        };
-
-        if let Some(id) = audit_id {
-            if let Some(gate) = crate::openhuman::approval::ApprovalGate::try_global() {
-                let exec = if outcome.is_ok() {
-                    crate::openhuman::approval::ExecutionOutcome::Success
-                } else {
-                    crate::openhuman::approval::ExecutionOutcome::Failure
-                };
-                gate.record_execution(
-                    &id,
-                    exec,
-                    outcome.as_ref().err().map(ToString::to_string).as_deref(),
-                );
-            }
-        }
-
-        outcome
-    }
-}
-
-/// [`CodeRunner`] adapter running sandboxed user code via
-/// `src/openhuman/sandbox/ops.rs` (`resolve_sandbox_policy` +
-/// `execute_in_sandbox`), modeled on
-/// `src/openhuman/tools/impl/system/node_exec.rs::run_sandboxed`.
-///
-/// **Mismatch handled here:** the sandbox runs a shell command string, not a
-/// `(language, source, input)` triple. `source` is treated as a function body
-/// receiving the serialized `input` items array and returning the node's
-/// output — this convention is a B1 design choice (not specified by the
-/// crate), matching the mock's "function body" tests
-/// (`tinyflows::nodes::integration::code` — e.g. `"source": "return 1;"`).
-///
-/// Requires `node`/`python3` on the `PATH` the sandbox backend runs under;
-/// there is no managed toolchain wiring here (unlike `node_exec`'s
-/// `NodeBootstrap`).
-///
-/// **Phase 2 — autonomy-tier gating:** a `code` node runs arbitrary user code
-/// in a sandbox, so it is treated as [`CommandClass::Write`] (state-changing but
-/// sandbox-bounded — not inherently catastrophic). Before dispatch it consults
-/// [`enforce_node_tier_gate`]: a read-only run `Block`s and never executes; a
-/// Supervised run then routes through the `ApprovalGate` (Write ⇒ `Prompt`); a
-/// Full run executes silently. This closes the prior gap where the code node had
-/// no policy check and no approval gate at all.
-pub struct OpenHumanCode {
-    pub config: Arc<Config>,
-    pub security: Arc<SecurityPolicy>,
-}
-
-const CODE_RUN_TIMEOUT_SECS: u64 = 60;
-
-#[async_trait]
-impl CodeRunner for OpenHumanCode {
-    async fn run(&self, language: CodeLanguage, source: &str, input: Value) -> Result<Value> {
-        // Autonomy-tier gate (Phase 2): sandboxed arbitrary-code execution is
-        // Write-class. A read-only run `Block`s here and never spawns anything;
-        // Supervised/Full fall through to the ApprovalGate below.
-        let tier_decision = enforce_node_tier_gate(&self.security, CommandClass::Write, "code")?;
-
-        // Approval gate (mirrors OpenHumanTools/OpenHumanHttp): `gate_call_for_tier`
-        // is what turns a Supervised-tier `Prompt` decision into a real human
-        // round-trip before any code runs — escalating past the flow's own
-        // `require_approval` toggle when the tier itself says "ask me" (Codex P1).
-        // A Deny short-circuits. The audit summary is computed on a redacted view
-        // of the request, never the raw source secrets, matching the other
-        // acting adapters.
-        let action = json!({ "language": format!("{language:?}"), "source": source });
-        let summary = crate::openhuman::approval::summarize_action("flows_code", &action);
-        let redacted = crate::openhuman::approval::redact_args(&action);
-        let (gate_outcome, audit_id) =
-            gate_call_for_tier(tier_decision, "flows_code", &summary, redacted).await;
-        if let crate::openhuman::approval::GateOutcome::Deny { reason } = gate_outcome {
-            return Err(EngineError::Capability(reason));
-        }
-
-        let outcome: Result<Value> = async {
-        let policy = resolve_sandbox_policy(
-            SandboxMode::Sandboxed,
-            &self.config.action_dir,
-            &self.config.runtime,
-            false,
-        );
-
-        // Work dir lives under `action_dir` (the sandbox workspace root). We keep
-        // its path *relative* to `action_dir` so the run command works on every
-        // backend: for Local, `execute_in_sandbox`'s `working_dir` is the host
-        // cwd; for Docker, `action_dir` is bind-mounted at `/workspace` with
-        // `-w /workspace`. Host-absolute paths would not exist inside the
-        // container, so we pass `action_dir` as the working dir and reference the
-        // script/input by their `action_dir`-relative paths.
-        let rel_dir = std::path::Path::new(".flows_code").join(uuid::Uuid::new_v4().to_string());
-        let work_dir = self.config.action_dir.join(&rel_dir);
-        tokio::fs::create_dir_all(&work_dir)
-            .await
-            .map_err(|e| EngineError::Capability(format!("failed to create code work dir: {e}")))?;
-
-        let (script_name, interpreter, script_body) = match language {
-            CodeLanguage::JavaScript => ("script.js", "node", js_harness(source)),
-            CodeLanguage::Python => ("script.py", "python3", python_harness(source)),
-        };
-        let script_path = work_dir.join(script_name);
-        let input_path = work_dir.join("input.json");
-
-        let input_json = serde_json::to_string(&input)
-            .map_err(|e| EngineError::Capability(format!("failed to serialize code input: {e}")))?;
-        tokio::fs::write(&script_path, script_body)
-            .await
-            .map_err(|e| EngineError::Capability(format!("failed to write code script: {e}")))?;
-        tokio::fs::write(&input_path, input_json)
-            .await
-            .map_err(|e| EngineError::Capability(format!("failed to write code input: {e}")))?;
-
-        // Backend-agnostic, `action_dir`-relative command paths (see above).
-        let rel_script = rel_dir.join(script_name);
-        let rel_input = rel_dir.join("input.json");
-        let command = format!(
-            "{} {} {}",
-            shell_quote(interpreter),
-            shell_quote(&rel_script.to_string_lossy()),
-            shell_quote(&rel_input.to_string_lossy()),
-        );
-
-        let mut extra_env = std::collections::HashMap::new();
-        if let Ok(host_path) = std::env::var("PATH") {
-            extra_env.insert("PATH".to_string(), host_path);
-        }
-
-        tracing::debug!(
-            target: "flows",
-            ?language,
-            work_dir = %work_dir.display(),
-            "[flows] code: running sandboxed script"
-        );
-
-        let exec_result = execute_in_sandbox(
-            &policy,
-            &command,
-            &self.config.action_dir,
-            extra_env,
-            std::time::Duration::from_secs(CODE_RUN_TIMEOUT_SECS),
-        )
-        .await;
-
-        // Always clean up the work dir — even when `execute_in_sandbox` itself
-        // errors (e.g. a spawn failure) — so temp scripts never leak.
-        if let Err(e) = tokio::fs::remove_dir_all(&work_dir).await {
-            tracing::debug!(target: "flows", error = %e, "[flows] code: failed to clean up work dir (non-fatal)");
-        }
-
-        let result = exec_result
-            .map_err(|e| EngineError::Capability(format!("sandbox execution failed: {e}")))?;
-
-        if !result.success() {
-            return Err(EngineError::Capability(format!(
-                "code node exited non-zero (timed_out={}): {}",
-                result.timed_out, result.stderr
-            )));
-        }
-
-        serde_json::from_str(result.stdout.trim())
-            .map_err(|e| EngineError::Capability(format!("code output was not valid JSON: {e}")))
-        }
-        .await;
-
-        // Close out the approval audit with the run's success/failure (mirrors
-        // OpenHumanTools/OpenHumanHttp).
-        if let Some(id) = audit_id {
-            if let Some(gate) = crate::openhuman::approval::ApprovalGate::try_global() {
-                let exec = if outcome.is_ok() {
-                    crate::openhuman::approval::ExecutionOutcome::Success
-                } else {
-                    crate::openhuman::approval::ExecutionOutcome::Failure
-                };
-                gate.record_execution(
-                    &id,
-                    exec,
-                    outcome.as_ref().err().map(ToString::to_string).as_deref(),
-                );
-            }
-        }
-
-        outcome
-    }
-}
-
-/// Wraps user `source` as a function body receiving `input`, executed by Node,
-/// printing the JSON result (or `null`) to stdout.
-fn js_harness(source: &str) -> String {
-    format!(
-        "const fs = require('fs');\n\
-         const input = JSON.parse(fs.readFileSync(process.argv[2], 'utf8'));\n\
-         const __result__ = (function(input) {{\n{source}\n}})(input);\n\
-         process.stdout.write(JSON.stringify(__result__ === undefined ? null : __result__));\n"
-    )
-}
-
-/// Wraps user `source` as a function body receiving `input`, executed by
-/// Python, printing the JSON result (or `null`) to stdout.
-fn python_harness(source: &str) -> String {
-    let indented: String = if source.trim().is_empty() {
-        "    pass".to_string()
-    } else {
-        source
-            .lines()
-            .map(|line| format!("    {line}"))
-            .collect::<Vec<_>>()
-            .join("\n")
-    };
-    format!(
-        "import sys, json\n\
-         with open(sys.argv[1]) as __f__:\n    input = json.load(__f__)\n\
-         def __user_fn__(input):\n{indented}\n    return None\n\
-         __result__ = __user_fn__(input)\n\
-         print(json.dumps(__result__))\n"
-    )
-}
-
-/// POSIX single-quote shell escaping, mirroring
-/// `tools/impl/system/node_exec.rs::shell_quote`.
-fn shell_quote(s: &str) -> String {
-    let escaped = s.replace('\'', "'\\''");
-    format!("'{escaped}'")
-}
-
-/// [`StateStore`] adapter over the `flows::` domain's `flow_state` KV table.
-pub struct FlowStateStore {
-    pub config: Arc<Config>,
-    pub namespace: String,
-}
-
-#[async_trait]
-impl StateStore for FlowStateStore {
-    async fn load(&self, key: &str) -> Result<Option<Value>> {
-        flows::kv_get(&self.config, &self.namespace, key)
-            .map_err(|e| EngineError::Capability(e.to_string()))
-    }
-
-    async fn store(&self, key: &str, value: Value) -> Result<()> {
-        flows::kv_set(&self.config, &self.namespace, key, &value)
-            .map_err(|e| EngineError::Capability(e.to_string()))
-    }
-}
-
-/// [`WorkflowResolver`] adapter over the `flows::` domain's saved-flow store.
-///
-/// A `sub_workflow` node that references a child by `workflow_id` (rather than
-/// embedding it inline) resolves through this adapter: the id is a saved flow's
-/// id, and [`flows::ops::load_flow_graph`] loads that flow's portable
-/// [`WorkflowGraph`] from the SQLite store. An unknown id maps to
-/// [`EngineError::Capability`], so the referencing node fails with a clear "no
-/// such workflow" error rather than silently no-op'ing.
-///
-/// The engine bounds recursion (its `MAX_SUB_WORKFLOW_DEPTH` depth counter) and
-/// rejects direct self-references before a child runs, so this adapter does not
-/// itself need cycle detection — it is a pure id → graph lookup.
-pub struct OpenHumanWorkflowResolver {
-    pub config: Arc<Config>,
-}
-
-#[async_trait]
-impl WorkflowResolver for OpenHumanWorkflowResolver {
-    async fn resolve(&self, workflow_id: &str) -> Result<WorkflowGraph> {
-        tracing::debug!(
-            target: "flows",
-            %workflow_id,
-            "[flows] sub_workflow resolver: resolving workflow_id to a saved flow graph"
-        );
-        match flows::ops::load_engine_compatible_flow_graph(&self.config, workflow_id) {
-            Ok(Some(graph)) => {
-                tracing::debug!(
-                    target: "flows",
-                    %workflow_id,
-                    node_count = graph.nodes.len(),
-                    "[flows] sub_workflow resolver: resolved saved flow graph"
-                );
-                Ok(graph)
-            }
-            Ok(None) => {
-                tracing::warn!(
-                    target: "flows",
-                    %workflow_id,
-                    "[flows] sub_workflow resolver: no saved flow with that workflow_id"
-                );
-                Err(EngineError::Capability(format!(
-                    "sub_workflow: no saved flow found for workflow_id '{workflow_id}'"
-                )))
-            }
-            Err(e) => {
-                tracing::error!(
-                    target: "flows",
-                    %workflow_id,
-                    error = %e,
-                    "[flows] sub_workflow resolver: failed to load saved flow graph"
-                );
-                Err(EngineError::Capability(format!(
-                    "sub_workflow: failed to load workflow_id '{workflow_id}': {e}"
-                )))
-            }
-        }
-    }
-}
 
 /// Builds the [`Capabilities`] bundle for one run, wiring each of the six
 /// host-injected traits to a real OpenHuman adapter (see each adapter above for
